# crypto/merkle: pre-allocate data slice in innherHash