--- a/CHANGELOG_PENDING.md
+++ b/CHANGELOG_PENDING.md
@@ -93,6 +93,7 @@ Friendly reminder: We have a [bug bounty program](https://hackerone.com/tendermi
 - [rpc] \#6265 set cache control in http-rpc response header (@JayT106)
 - [statesync] \#6378 Retry requests for snapshots and add a minimum discovery time (5s) for new snapshots.
 - [node/state] \#6370 graceful shutdown in the consensus reactor (@JayT106)
+- [crypto/merkle] \#6443 Improve HashAlternatives performance (@cuonglm)
 
 ### BUG FIXES