So we can reduce pressure on runtime for checking that slice has enough
capacity before appending.

Benchmark result shows less memory usage and faster runtime:

```
name                          old time/op    new time/op    delta
HashAlternatives/recursive-8    25.2µs ± 1%    19.6µs ± 1%  -22.07%  (p=0.000 n=10+9)
HashAlternatives/iterative-8    24.2µs ± 1%    19.0µs ± 0%  -21.28%  (p=0.000 n=10+10)

name                          old alloc/op   new alloc/op   delta
HashAlternatives/recursive-8    25.4kB ± 0%    19.1kB ± 0%  -24.92%  (p=0.000 n=10+10)
HashAlternatives/iterative-8    28.1kB ± 0%    21.8kB ± 0%  -22.54%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
HashAlternatives/recursive-8       497 ± 0%       398 ± 0%  -19.92%  (p=0.000 n=10+10)
HashAlternatives/iterative-8       498 ± 0%       399 ± 0%  -19.88%  (p=0.000 n=10+10)
```

Fixes #6442